requirements.txt
pytest
loguru
playwright
opencv-python
numpy
pyautogui
robotframework
pytest-playwright